Cont. implementing new model.    - simple baseline layer

+++ gnn_models/mip_bipartite_class.py
@@ -276,7 +276,7 @@ class SimpleNet(torch.nn.Module):
 
         var = self.var_assigment_4(var_node_features_4)
         print(var.mean())
-        print(data.y.mean())
+        print(data.y.float().mean())
         #print(err_1.min(), print(err_1.max()))
 
         x = torch.cat(
